Make goonline button a fab
