+++ src/components/pages/layouts/MainLayout.tsx
@@ -6,13 +6,16 @@ import Banner from '../../shared/Banner'
 import OfflineNotification from '../../shared/OfflineNotification'
 import OnlineNotificationSnackbar from '../../shared/OnlineNotificationSnackbar'
 import ResponsiveNavBar from '../../shared/navbar/ResponsiveAppBar'
+import GoOnlineButton from '../../shared/GoOnlineButton'
 
 const MainLayout = () => {
   const { isOffline } = useConnectionStatus()
 
   const offlineBanner = isOffline() ? <Banner bannerContent={<OfflineNotification/>}/> : null
 
-  return (<>
+  return (
+    <>
+      <GoOnlineButton/>
       <ResponsiveNavBar/>
       {offlineBanner}
       <Outlet/>
